Enum zng_wgt_window::SaveState

pub enum SaveState {
    Enabled {
        key: Option<ConfigKey>,
        loading_timeout: Duration,
    },
    Disabled,
}

Window persistence config.

See the save_state property for more details.

Variants

Enabled

Fields

key: Option<ConfigKey>

Config key that identifies the window.

If None a key is generated for the window, using the window_key method.

loading_timeout: Duration

Maximum time to keep the window in the loading state awaiting for the config to load.

If the config fails to load in this time frame the window is opened in it’s default state.

This is one second by default.

Save and restore state.

Disabled

Don’t save nor restore state.

Implementations

impl SaveState

pub fn enabled() -> Self

Default, enabled, no key, delay 1s.

pub fn window_key(&self, id: WindowId) -> Option<ConfigKey>

Gets the config key used for the window identified by id.

pub fn loading_timeout(&self) -> Option<Duration>

Get the loading timeout if it is enabled and the duration is greater than zero.

pub fn is_enabled(&self) -> bool

Returns true if it is enabled.