zng_wgt_window::events

Function on_window_load

pub fn on_window_load(
    child: impl IntoUiNode,
    handler: Handler<WindowOpenArgs>,
) -> UiNode

P On window loaded.

This event notifies once per window, after the first layout and all WindowLoadingHandle have expired or dropped.

Preview

You can preview this event using on_pre_window_load. Otherwise the handler is only called after the widget content has a chance to stop propagation.

Async

You can use async event handlers with this property.