zng_wgt_window::events

Function on_window_state_changed

pub fn on_window_state_changed(
    child: impl UiNode,
    handler: impl WidgetHandler<WindowChangedArgs>,
) -> impl UiNode

P On window state changed.

This event notifies every time the user or the app changes the WindowVars::state.

Preview

You can preview this event using on_pre_window_state_changed. Otherwise the handler is only called after the widget content has a chance to stop propagation.

Async

You can use async event handlers with this property.