zng_wgt_window

Function headless_monitor

pub fn headless_monitor(monitor: impl IntoValue<HeadlessMonitor>)

c Imaginary monitor used by the window when it runs in headless mode.

Capture-Only

This property is capture-only, it only defines a property signature, it does not implement any behavior by itself. Widgets can capture and implement this property as part of their intrinsics, otherwise it will have no effect if set on a widget that does not implement it.