Function zng_wgt_window::render_mode

pub fn render_mode(mode: impl IntoValue<Option<RenderMode>>)

c Render performance mode overwrite for this window, if set to None the WINDOWS.default_render_mode is used.

The view-process will try to match the mode, if it is not available a fallback mode is selected, see RenderMode for more details about each mode and fallbacks.

Capture-Only

This property is capture-only, it only defines a property signature, it does not implement any behavior by itself. Widgets can capture and implement this property as part of their intrinsics, otherwise it will have no effect if set on a widget that does not implement it.