zng_wgt_window

Function save_state

pub fn save_state(
    child: impl IntoUiNode,
    enabled: impl IntoValue<SaveState>,
) -> UiNode

P Save and restore the window state.

If enabled a config entry is created for the window state in CONFIG, and if a config backend is set the window state is persisted on change and restored when the app reopens.

This property is enabled by default in the Window! widget, without a key. Note that without a config key the state only actually enables if the window root widget ID or the window ID have a name.