zng_wgt_window

Struct DefaultStyle

pub struct DefaultStyle(/* private fields */);

W Default window style.

See also register_style_fn for how to set a style for all windows in the app.

Implementations

impl DefaultStyle

pub fn position(&self, position: impl IntoVar<Point>)

P Binds the position window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn monitor(&self, monitor: impl IntoVar<MonitorQuery>)

P Binds the monitor window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn state(&self, state: impl IntoVar<WindowState>)

P Binds the state window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn size(&self, size: impl IntoVar<Size>)

P Binds the size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn min_size(&self, min_size: impl IntoVar<Size>)

P Binds the min_size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn max_size(&self, max_size: impl IntoVar<Size>)

P Binds the max_size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn font_size(&self, font_size: impl IntoVar<Length>)

P Binds the font_size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn chrome(&self, chrome: impl IntoVar<bool>)

P Binds the chrome window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn icon(&self, icon: impl IntoVar<WindowIcon>)

P Binds the icon window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn title(&self, title: impl IntoVar<Txt>)

P Binds the title window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn auto_size(&self, auto_size: impl IntoVar<AutoSize>)

P Binds the auto_size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn auto_size_origin(&self, auto_size_origin: impl IntoVar<Point>)

P Binds the auto_size_origin window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn resizable(&self, resizable: impl IntoVar<bool>)

P Binds the resizable window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn movable(&self, movable: impl IntoVar<bool>)

P Binds the movable window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn always_on_top(&self, always_on_top: impl IntoVar<bool>)

P Binds the always_on_top window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn visible(&self, visible: impl IntoVar<bool>)

P Binds the visible window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn taskbar_visible(&self, taskbar_visible: impl IntoVar<bool>)

P Binds the taskbar_visible window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn parent(&self, parent: impl IntoVar<Option<WindowId>>)

P Binds the parent window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn modal(&self, modal: impl IntoVar<bool>)

P Binds the modal window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn color_scheme(&self, color_scheme: impl IntoVar<Option<ColorScheme>>)

P Binds the color_scheme window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn accent_color(&self, accent_color: impl IntoVar<Option<LightDark>>)

P Binds the accent_color window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn enabled_buttons(&self, enabled_buttons: impl IntoVar<WindowButton>)

P Binds the enabled_buttons window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn parallel(&self, parallel: impl IntoVar<Parallel>)

P Binds the parallel window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn frame_capture_mode( &self, frame_capture_mode: impl IntoVar<FrameCaptureMode>, )

P Binds the frame_capture_mode window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn x(&self, x: impl IntoVar<Length>)

P Binds the x of the position window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn y(&self, y: impl IntoVar<Length>)

P Binds the y of the position window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn width(&self, width: impl IntoVar<Length>)

P Binds the width of the size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn height(&self, height: impl IntoVar<Length>)

P Binds the height of the size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn min_width(&self, min_width: impl IntoVar<Length>)

P Binds the width of the min_size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn min_height(&self, min_height: impl IntoVar<Length>)

P Binds the height of the min_size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn max_width(&self, max_width: impl IntoVar<Length>)

P Binds the width of the max_size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn max_height(&self, max_height: impl IntoVar<Length>)

P Binds the height of the max_size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl DefaultStyle

pub fn clear_color(&self, color: impl IntoVar<Rgba>)

P Window clear color.

Color used to clear the previous frame pixels before rendering a new frame. It is visible if window content does not completely fill the content area, this can happen if you do not set a background or the background is semi-transparent, also can happen during very fast resizes.

impl DefaultStyle

pub fn prefer_custom_chrome(&self, prefer: impl IntoVar<bool>)

P Gets if WINDOWS.system_chrome prefers custom chrome.

Note that you must set chrome to false when using this to provide a custom chrome.

impl DefaultStyle

pub fn custom_chrome_adorner_fn( &self, custom_chrome: impl IntoVar<WidgetFn<()>>, )

P Adorner property specific for custom chrome overlays.

This property behaves exactly like adorner_fn. Using it instead of adorner frees the adorner property for other usage in the window instance or in derived window types.

Note that you can also set the custom_chrome_padding_fn to ensure that the content is not hidden behind the adorner.

impl DefaultStyle

pub fn custom_chrome_padding_fn(&self, padding: impl IntoVar<SideOffsets>)

P Extra padding for window content in windows that display a custom_chrome_adorner_fn.

impl DefaultStyle

pub fn widget_new() -> Self

Start building a new instance.

pub fn widget_type() -> WidgetType

Gets the widget type info.

impl DefaultStyle

pub fn safe_padding(&self, padding: impl IntoVar<SideOffsets>)

P Padding required to avoid physical screen obstructions.

By default this is WINDOW.vars().safe_padding() that is defined by the operating system. You can unset this property to implement your own unsafe area handling.

impl DefaultStyle

pub fn start_position(&self, position: impl IntoValue<StartPosition>)

B Defines how the window is positioned when it first opens.

impl DefaultStyle

pub fn allow_transparency(&self, allow: impl IntoValue<bool>)

B If semi-transparent content is see-through, mixing with the operating system pixels behind the window.

Note that to actually see behind the window you must set the clear_color and background_color to a transparent color. The composition is a simple alpha blend, effects like blur do not apply to the pixels behind the window.