Crate zng_wgt_window

Window widget, properties, properties and nodes.

Crate

This crate is part of the zng project.

Cargo Features

This crate provides 2 feature flags, 1 enabled by default.

"debug_default"

Enable the zng-wgt crate debug defaults.

Enabled by default.

"dyn_closure"

Enable the zng-wgt crate dynamic closures.

Modules

events

Window events, on_window_open and on_window_close_requested.

Structs

Window

W A window container.

Enums

BlockWindowLoad

Defines if a widget load affects the parent window load.

SaveState

Window or widget persistence config.

Functions

accent_color

P Binds the accent_color window var with the property value.

allow_transparency

c If semi-transparent content is see-through, mixing with the operating system pixels behind the window.

always_on_top

P Binds the always_on_top window var with the property value.

auto_size

P Binds the auto_size window var with the property value.

auto_size_origin

P Binds the auto_size_origin window var with the property value.

chrome

P Binds the chrome window var with the property value.

clear_color

P Window clear color.

color_scheme

P Binds the color_scheme window var with the property value.

config_block_window_load

P Block window load until CONFIG.status is idle.

custom_chrome_adorner_fn

P Adorner property specific for custom chrome overlays.

custom_chrome_padding_fn

P Extra padding for window content in windows that display a custom_chrome_adorner_fn.

enabled_buttons

P Binds the enabled_buttons window var with the property value.

fallback_chrome

Custom window chrome adorner used when the window manager does not provide one.

font_size

P Binds the font_size window var with the property value.

frame_capture_mode

P Binds the frame_capture_mode window var with the property value.

headless_monitor

c Imaginary monitor used by the window when it runs in headless mode.

height

P Binds the height of the size window var with the property value.

icon

P Binds the icon window var with the property value.

kiosk

c Lock-in kiosk mode.

max_height

P Binds the height of the max_size window var with the property value.

max_size

P Binds the max_size window var with the property value.

max_width

P Binds the width of the max_size window var with the property value.

min_height

P Binds the height of the min_size window var with the property value.

min_size

P Binds the min_size window var with the property value.

min_width

P Binds the width of the min_size window var with the property value.

modal

P Binds the modal window var with the property value.

monitor

P Binds the monitor window var with the property value.

movable

P Binds the movable window var with the property value.

needs_fallback_chrome

P Gets if is not headless, chrome is true, state is not fullscreen but WINDOWS.system_chrome reports the system does not provide window decorations.

on_close

P On window close.

on_close_requested

P On window close requested.

on_exited_fullscreen

P On window is no longer fullscreen.

on_frame_image_ready

P On window frame rendered.

on_fullscreen

P On window enter one of the fullscreen states.

on_load

P Event just after the window loads.

on_maximized

P On window maximized.

on_minimized

P On window minimized.

on_moved

P On window position changed.

on_open

P Event just after the window opens.

on_resized

P On window size changed.

on_restored

P On window state changed to Normal.

on_state_changed

P On window state changed.

on_unmaximized

P On window exited the maximized state.

on_unminimized

P On window exited the minimized state.

parent

P Binds the parent window var with the property value.

position

P Binds the position window var with the property value.

prefer_custom_chrome

P Gets if WINDOWS.system_chrome prefers custom chrome.

render_mode

c Render performance mode overwrite for this window, if set to None the WINDOWS.default_render_mode is used.

resizable

P Binds the resizable window var with the property value.

save_state

P Save and restore the window state.

save_state_node

Helper node for implementing widgets save.

size

P Binds the size window var with the property value.

start_focused

c If the window is steals keyboard focus on open.

start_position

c Defines how the window is positioned when it first opens.

state

P Binds the state window var with the property value.

taskbar_visible

P Binds the taskbar_visible window var with the property value.

title

P Binds the title window var with the property value.

visible

P Binds the visible window var with the property value.

width

P Binds the width of the size window var with the property value.

x

P Binds the x of the position window var with the property value.

y

P Binds the y of the position window var with the property value.