zng_wgt_window
Module events

Window events, on_window_open and on_window_close_requested.

These events are re-exported by Window! as on_open and on_close_requested, but you can attach then in any widget inside a window using the property full name.

There is no event property for the WINDOW_CLOSE_EVENT because that event notifies after the window is deinited. You can use on_deinit in the Window! widget to handle a window closing, or create an app level handler for the event using EVENTS.

Functions

on_frame_image_ready

P On window frame rendered.

on_ime

P On Input Method Editor event.

on_pre_frame_image_ready

P Preview on_frame_image_ready event.

on_pre_ime

P Preview on_ime event.

on_pre_window_changed

P Preview on_window_changed event.

on_pre_window_close

P Preview on_window_close event.

on_pre_window_close_requested

P Preview on_window_close_requested event.

on_pre_window_exited_fullscreen

P Preview on_window_exited_fullscreen event.

on_pre_window_fullscreen

P Preview on_window_fullscreen event.

on_pre_window_load

P Preview on_window_load event.

on_pre_window_maximized

P Preview on_window_maximized event.

on_pre_window_minimized

P Preview on_window_minimized event.

on_pre_window_moved

P Preview on_window_moved event.

on_pre_window_open

P Preview on_window_open event.

on_pre_window_resized

P Preview on_window_resized event.

on_pre_window_restored

P Preview on_window_restored event.

on_pre_window_state_changed

P Preview on_window_state_changed event.

on_pre_window_unmaximized

P Preview on_window_unmaximized event.

on_pre_window_unminimized

P Preview on_window_unminimized event.

on_window_changed

P On window moved, resized or other state changed.

on_window_close

P On window closed.

on_window_close_requested

P On window close requested.

on_window_exited_fullscreen

P On window state changed from WindowState::is_fullscreen.

on_window_fullscreen

P On window state changed to WindowState::is_fullscreen.

on_window_load

P On window loaded.

on_window_maximized

P On window state changed to WindowState::Maximized.

on_window_minimized

P On window state changed to WindowState::Minimized.

on_window_moved

P On window position changed.

on_window_open

P On window opened.

on_window_resized

P On window size changed.

on_window_restored

P On window state changed to WindowState::Normal.

on_window_state_changed

P On window state changed.

on_window_unmaximized

P On window state changed from WindowState::Maximized.

on_window_unminimized

P On window state changed from WindowState::Minimized.