zng_wgt_window

Trait WINDOWS_Ext

pub trait WINDOWS_Ext {
    // Required method
    fn register_style_fn(&self, style_fn: impl IntoVar<StyleFn>);
}

Extension methods for WINDOWS.

Required Methods

fn register_style_fn(&self, style_fn: impl IntoVar<StyleFn>)

Set the style_fn in all windows instantiated after this call.

This method is the recommended entry point for themes. It uses register_root_extender to inject the style in every new window instance.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.